feat: 퀴즈 생성도 동일하게 조정
